Assignment4/requirements.txt
langchain
langchain-community
langchain-core
pymongo
python-dotenv
requests
langchain-openai
